feat: cleanup group list and side menu components
