+++ frontend/src/components/side-menu/group-list/index.tsx
@@ -1,19 +1,18 @@
 import { JSX } from 'solid-js/jsx-runtime';
 import { styled } from '../../../utils';
-import { SideMenuIconContent } from '../icon-content';
-import { SideMenuItemContainer } from '../item-container';
 import ExpandMoreSvg from './icons/expand_more.svg';
-import { countText, expandMoreIcon, header, itemList, name } from './index.css';
-import { ParentProps, Show, children, createSignal } from 'solid-js';
+import { container, countText, expandMoreIcon, header, iconBox, name } from './index.css';
+import { ParentProps, Show, createSignal } from 'solid-js';
 
+const Container = styled('div', container);
 const Header = styled('div', header);
 const Name = styled('span', name);
 const CountText = styled('span', countText);
-const ItemList = styled('ul', itemList);
 const ExpandMoreIcon = styled(ExpandMoreSvg, expandMoreIcon);
+const IconBox = styled('div', iconBox);
 
 export type SideMenuGroupListProp = ParentProps<{
-  icon: JSX.Element;
+  icon?: JSX.Element;
   itemCount: number;
   name?: string;
 
@@ -23,27 +22,23 @@ export type SideMenuGroupListProp = ParentProps<{
 export const SideMenuGroupList = (props: SideMenuGroupListProp) => {
   const [expanded, setExpanded] = createSignal(!!props.defaultExpanded);
 
-  const childList = children(() => props.children);
-  const childrenCount = () => {
-    const target = childList();
-
-    return Array.isArray(target) ? target.length : Number(!!childList());
-  };
-
-  return <SideMenuItemContainer>
+  return <Container>
     <Header onClick={() => setExpanded(!expanded())}>
-      <SideMenuIconContent icon={props.icon}>
-        <Show when={props.name}>
-          <Name>{props.name}</Name>
-        </Show>
-        <Show when={props.itemCount > 0}>
-          <CountText>{props.itemCount}</CountText>
-        </Show>
-        <ExpandMoreIcon data-expanded={expanded()} />
-      </SideMenuIconContent>
+      <Show when={props.icon}>
+        <IconBox>
+          {props.icon}
+        </IconBox>
+      </Show>
+      <Show when={props.name}>
+        <Name>{props.name}</Name>
+      </Show>
+      <Show when={props.itemCount > 0}>
+        <CountText>{props.itemCount}</CountText>
+      </Show>
+      <ExpandMoreIcon data-expanded={expanded()} />
     </Header>
-    <Show when={expanded() && childrenCount() > 0}>
-      <ItemList>{childList()}</ItemList>
+    <Show when={expanded()}>
+      {props.children}
     </Show>
-  </SideMenuItemContainer>;
+  </Container>;
 };
