+++ frontend/src/components/side-menu/index.tsx
@@ -1,13 +1,12 @@
 import { For, JSX, ParentProps, Show, children } from 'solid-js';
 import { styled } from '../../utils';
-import { container, contentItem, contentList, head, headContainer, name } from './index.css';
+import { container, contentList, head, headContainer, name } from './index.css';
 
 const Container = styled('div', container);
 const Head = styled('div', head);
 const Name = styled('span', name);
 const HeadContainer = styled('div', headContainer);
 const ContentList = styled('ul', contentList);
-const ContentItem = styled('li', contentItem);
 
 export type SideMenuProp = ParentProps<{
   name: string,
@@ -35,7 +34,7 @@ export const SideMenu = (props: SideMenuProp) => {
     <Show when={childList().length > 0}>
       <ContentList>
         <For each={childList()}>
-          {(child) => <ContentItem>{child}</ContentItem>}
+          {(child) => <li>{child}</li>}
         </For>
       </ContentList>
     </Show>
