--- frontend/src/components/side-menu/item-container.css.ts
@@ -1,8 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const sideMenuItemContainer = style({
-  backgroundColor: 'rgba(255, 255, 255, 0.5)',
-
-  border: '1px solid #DFDEE0',
-  borderRadius: '0.5rem',
-});
