--- frontend/src/components/side-menu/item-container.stories.tsx
@@ -1,16 +1,0 @@
-import { StoryFn } from 'storybook-solidjs';
-import { SideMenuItemContainer } from './item-container';
-
-export default {
-  title: 'KiwiTalk/SideMenu/ItemContainer',
-  component: SideMenuItemContainer,
-};
-
-const Template: StoryFn<typeof SideMenuItemContainer> = (args) => {
-  return <SideMenuItemContainer {...args}>
-    Example Text
-  </SideMenuItemContainer>;
-};
-
-export const Default = Template.bind({});
-Default.args = {};
