--- frontend/src/components/side-menu/item-container.tsx
@@ -1,4 +1,0 @@
-import { styled } from '../../utils';
-import { sideMenuItemContainer } from './item-container.css';
-
-export const SideMenuItemContainer = styled('div', sideMenuItemContainer);
